+++ api/controllers/user.js
@@ -227,21 +227,6 @@ module.exports = {
     }
   },
 
-  async socialAuthlogin(req, res, next) {
-    const { user } = req;
-
-    try {
-      const token = await jwt.generateToken(user.dataValues);
-
-      return response.success(res, 200, {
-        message: `${user.dataValues.email} successfully logged in.`,
-        token
-      });
-    } catch (error) {
-      return next({ message: `${error.message}` });
-    }
-  },
-
   async uploadUserImage(req, res, next) {
     const { params, file } = req;
     console.log(file);
